added lib.rs for truly library experience
